Improve Remote Server logging and fix wrong types in online API specification
Add capability to log the client IP address in the server log against each transaction. For backward compatibility, this is defaulted off and can be enabled through the Server setup screen.
Most lines in the server log now include the method name to make following a transaction in the log easier.
The ClientID, ClientTransactionID and ServerTransactionID fields in the log are now right aligned.
The range of ClientIDs that are auto generated by Remote Clients has been reduced to to 1::65536 so that they fit better in the allocated log space.
Some incorrect PUT parameter types in the online API documentation have been fixed e.g. some integer values were documented as being strings, the API itself has not changed.
Extensive refactoring to store each transaction's information in an object and to pass this to methods rather than passing in many individual transaction attributes.