Replace `default` with `example` in OpenAPI
`default` implies a default value that will be used if parameter is not provided, but the parameters are all mandatory and values like `ClientID=123` are certainly not defaults.

Instead, a better field to use is `example` which still sets this value for input fields in the Swagger UI but removes the "Default value: ..." text.

It's also a better hint for tooling so that it doesn't autogenerate functions with optional parameters set to these values.